+++ core/src/main/java/de/hipphampel/validation/core/report/BooleanReporter.java
@@ -22,6 +22,7 @@
  */
 package de.hipphampel.validation.core.report;
 
+import de.hipphampel.validation.core.execution.ValidationContext;
 import de.hipphampel.validation.core.path.Path;
 import de.hipphampel.validation.core.rule.Result;
 import de.hipphampel.validation.core.rule.Rule;
@@ -74,7 +75,7 @@ public class BooleanReporter implements Reporter<Boolean> {
   }
 
   @Override
-  public void add(Path path, Object facts, Rule<?> rule, Result result) {
+  public void add(ValidationContext context, Object facts, Path path, Rule<?> rule, Result result) {
     report.compareAndSet(true, skippedIsFailed ? result.isOk() : !result.isFailed());
   }